+++ dmr/sse/metadata.py
@@ -20,6 +20,7 @@ from dmr.negotiation import ContentType
 from dmr.openapi import OpenAPIContext
 from dmr.openapi.objects import Response
 from dmr.serializer import BaseSerializer
+from dmr.sse.validation import check_event_field
 
 _DataT_co = TypeVar('_DataT_co', covariant=True)
 
@@ -173,6 +174,10 @@ class SSEvent(Generic[_DataT_co]):
             and comment is None
         ):
             raise ValueError('At least one event field must be non-None')
+
+        # Check null byte and new lines:
+        check_event_field(id, field_name='id')
+        check_event_field(event, field_name='event')
 
         self.data = data
         self.event = event